refactor: move 'AnimatedSpan' to related hook
