+++ src/hooks/useAnimatedChildren/index.ts
@@ -1,1 +1,2 @@
+export { AnimatedSpan } from './AnimatedSpan';
 export { useAnimatedChildren } from './useAnimatedChildren';
